+++ src/main/kotlin/org/codefreak/codefreak/service/AdministrationService.kt
@@ -27,4 +27,6 @@ class AdministrationService : BaseService() {
   fun getMaxUploadSize(): Long = multipartConfigElement.maxFileSize
 
   fun getDefaultIdeDockerImageName(): String = config.ide.image
+
+  fun getDefaultEvaluationTimeout(): Long = config.evaluation.defaultTimeout
 }